refactor(Utils): ♻️ Modernized timer logic & moved -> composables.     -   Moved `startTimer` as a new composable utility, which improves     code organization by moving logic closer to where it is reactive and     functional. Renamed `startTimer` -> `useCarousel`**, which more     accurately reflects its purpose and aligns with standard Vue naming     conventions for reusable logic and enhanced hook flexibility, which     allows the carousel logic to be reused across different components     with various timing and transition requirements.
    -   Removed legacy utilities which involved removing the redundant
    `next` function to reduce technical garbage and clean the codebase.

    -   Optimized state management, which provides a more robust and
    testable foundation for time-based UI transitions.

    -   Extracted `mapTimeline` and `mapReferences` functions,
    by separates data transformation logic from the core preprocessor to
    improve modularity and testability.

    -   simplified the preprocessing pipeline, which enhances code
    readability by allowing the main preprocessor to focus on orchestration
    rather than low-level mapping.

    -   Improved architectural separation, which provides a cleaner
    foundation for future data integrations and reduces the risk of side
    effects when updating mapping rules.
