+++ frontend/utils/utils.ts
@@ -8,40 +8,3 @@ export function generateHexID()
 
     return result;
 }
-
-export function next(arr: Array<Record<string, any>>)
-    {
-
-        const n = arr.length;
-        if (n <= 1) return;
-
-        // Iterate through the array
-        for (let i = 0; i < n; i++)
-        {
-            const length = n - 1;
-            const source = arr[i];
-
-            const j = i === length ? 0 : i + 1;
-            const target = arr[j];
-
-            //  Swap the visibility
-            if (source.isVisible)
-            {
-                const temp: boolean = source.isVisible;
-                source.isVisible = target.isVisible;
-                target.isVisible = temp;
-                break;
-            }
-
-        }
-        //console.log(arr);
-    };
-
-export function startTimer(arr: Array<Record<string, any>>)
-    {
-        //  Set the interval
-        const n = 10000;
-
-        //  Start the timer
-        return setInterval(() => next(arr), n);
-    };
